Minor refactoring of some internal usage classes.
