+++ mlrose/__init__.py
@@ -12,7 +12,6 @@ from .algorithms.mimic import (mimic)
 from .algorithms.decay import GeomDecay, ArithDecay, ExpDecay, CustomSchedule
 from .algorithms.crossovers import OnePointCrossOver, UniformCrossOver, TSPCrossOver
 from .algorithms.mutators import ChangeOneMutator, DiscreteMutator, SwapMutator, ShiftOneMutator
-from .algorithms.decorators import (get_short_name)
 from .fitness import (OneMax, FlipFlop, FourPeaks, SixPeaks, ContinuousPeaks,
                       Knapsack, TravellingSales, Queens, MaxKColor, 
                       CustomFitness)
