+++ mlrose/algorithms/mimic.py
@@ -6,7 +6,7 @@
 
 import numpy as np
 
-from mlrose.algorithms.decorators import short_name
+from mlrose.decorators import short_name
 
 
 @short_name('mimic')
