+++ mlrose/decorators/short_name_decorator.py
